requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
cffi==1.15.1
charset-normalizer==2.1.1 
click==8.1.3
flake8==6.0.0
frozenlist==1.3.3
gTTS==2.3.0
idna==3.4
magic-filter==1.0.9
mccabe==0.7.0
multidict==6.0.4
openai==0.27.8
pycodestyle==2.10.0
pycparser==2.21
pydub==0.25.1
pyflakes==3.0.1
pyTelegramBotAPI==4.12.0
python-dotenv==0.21.0
pyttsx3==2.90
requests==2.28.1
schedule==1.2.0
six==1.16.0
sounddevice==0.4.6
SpeechRecognition==3.9.0
tqdm==4.65.0
urllib3==1.26.13
yarl==1.8.2
